Modified and moved randon.js to api/discover
